+++ src/window.h
@@ -12,8 +12,8 @@ private:
     HWND hwnd;
     HINSTANCE hInstance;
     static const char CLASS_NAME[];
-    static int window_width;
-    static int window_height;
+    static size_t window_width;
+    static size_t window_height;
     static std::mutex agentMutex;
     static LRESULT CALLBACK WindowProc(HWND hwnd, UINT uMsg, WPARAM wParam, LPARAM lParam);
 };
